+++ rust/crd/src/discovery.rs
@@ -1,4 +1,4 @@
-use stackable_operator::utils::cluster_domain::KUBERNETES_CLUSTER_DOMAIN;
+use stackable_operator::utils::cluster_info::KubernetesClusterInfo;
 
 use crate::{HTTPS_PORT, HTTP_PORT};
 
@@ -11,13 +11,13 @@ pub struct TrinoPodRef {
 }
 
 impl TrinoPodRef {
-    pub fn fqdn(&self) -> String {
-        let cluster_domain = KUBERNETES_CLUSTER_DOMAIN
-            .get()
-            .expect("KUBERNETES_CLUSTER_DOMAIN must first be set by calling initialize_operator");
+    pub fn fqdn(&self, cluster_info: &KubernetesClusterInfo) -> String {
         format!(
-            "{}.{}.{}.svc.{}",
-            self.pod_name, self.role_group_service_name, self.namespace, cluster_domain
+            "{pod_name}.{service_name}.{namespace}.svc.{cluster_domain}",
+            pod_name = self.pod_name,
+            service_name = self.role_group_service_name,
+            namespace = self.namespace,
+            cluster_domain = cluster_info.cluster_domain
         )
     }
 }
@@ -36,11 +36,11 @@ impl TrinoDiscovery {
         }
     }
 
-    pub fn discovery_uri(&self) -> String {
+    pub fn discovery_uri(&self, cluster_info: &KubernetesClusterInfo) -> String {
         format!(
             "{}://{}:{}",
             self.protocol,
-            self.pod_ref.fqdn(),
+            self.pod_ref.fqdn(cluster_info),
             self.protocol.port()
         )
     }