+++ rust/operator-binary/src/main.rs
@@ -51,6 +51,7 @@ async fn main() -> anyhow::Result<()> {
             product_config,
             watch_namespace,
             tracing_target,
+            cluster_info_opts,
         }) => {
             stackable_operator::logging::initialize_logging(
                 "TRINO_OPERATOR_LOG",
@@ -70,9 +71,11 @@ async fn main() -> anyhow::Result<()> {
                 "/etc/stackable/trino-operator/config-spec/properties.yaml",
             ])?;
 
-            let client =
-                stackable_operator::client::initialize_operator(Some(OPERATOR_NAME.to_string()))
-                    .await?;
+            let client = stackable_operator::client::initialize_operator(
+                Some(OPERATOR_NAME.to_string()),
+                &cluster_info_opts,
+            )
+            .await?;
 
             let cluster_controller = Controller::new(
                 watch_namespace.get_api::<TrinoCluster>(&client),