+++ CHANGELOG.md
@@ -7,9 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added support for Trino 455 ([#638]).
-- The operator can now run on Kubernetes clusters using a non-default cluster domain. It should automatically detect the
-  correct domain to use, but you can also use the env var `KUBERNETES_CLUSTER_DOMAIN` to set the domain explicitly
-  or use the helm-chart property `kubernetesClusterDomain` ([#655]).
+- The operator can now run on Kubernetes clusters using a non-default cluster domain.
+  Use the env var `KUBERNETES_CLUSTER_DOMAIN` or the operator Helm chart property `kubernetesClusterDomain` to set a non-default cluster domain ([#655]).
 
 ### Changed